refactor: separate ActionButton into components
