+++ packages/sdk/src/react/ui/components/_internals/action-button/components/NonOwnerActions.tsx
@@ -1,0 +1,72 @@
+import type { Hex } from 'viem';
+import type { Order, OrderbookKind } from '../../../../../_internal';
+import { CollectibleCardAction } from '../types';
+import { ActionButtonBody } from './ActionButtonBody';
+import { useBuyModal } from '../../../../modals/BuyModal';
+import { useMakeOfferModal } from '../../../../modals/MakeOfferModal';
+import { InvalidStepError } from '../../../../../../utils/_internal/error/transaction';
+import SvgCartIcon from '../../../../icons/CartIcon';
+
+type NonOwnerActionsProps = {
+	action: CollectibleCardAction;
+	tokenId: string;
+	collectionAddress: Hex;
+	chainId: string;
+	orderbookKind?: OrderbookKind;
+	lowestListing?: Order;
+};
+
+export function NonOwnerActions({
+	action,
+	tokenId,
+	collectionAddress,
+	chainId,
+	orderbookKind,
+	lowestListing,
+}: NonOwnerActionsProps) {
+	const { show: showBuyModal } = useBuyModal();
+	const { show: showMakeOfferModal } = useMakeOfferModal();
+
+	if (action === CollectibleCardAction.BUY) {
+		if (!lowestListing) {
+			throw new InvalidStepError('BUY', 'lowestListing is required');
+		}
+
+		return (
+			<ActionButtonBody
+				action={CollectibleCardAction.BUY}
+				tokenId={tokenId}
+				label="Buy now"
+				onClick={() =>
+					showBuyModal({
+						collectionAddress,
+						chainId,
+						tokenId,
+						order: lowestListing,
+					})
+				}
+				icon={SvgCartIcon}
+			/>
+		);
+	}
+
+	if (action === CollectibleCardAction.OFFER) {
+		return (
+			<ActionButtonBody
+				action={CollectibleCardAction.OFFER}
+				tokenId={tokenId}
+				label="Make an offer"
+				onClick={() =>
+					showMakeOfferModal({
+						collectionAddress,
+						chainId,
+						collectibleId: tokenId,
+						orderbookKind,
+					})
+				}
+			/>
+		);
+	}
+
+	return null;
+}
