+++ packages/sdk/src/react/ui/components/_internals/action-button/hooks/useActionButtonLogic.ts
@@ -1,0 +1,84 @@
+import { useEffect } from 'react';
+import { useAccount } from 'wagmi';
+import { CollectibleCardAction } from '../types';
+import { actionButtonStore, executePendingActionIfExists } from '../store';
+
+type UseActionButtonLogicProps = {
+	tokenId: string;
+	owned?: boolean;
+	action: CollectibleCardAction;
+	onCannotPerformAction?: (
+		action: CollectibleCardAction.BUY | CollectibleCardAction.OFFER,
+	) => void;
+};
+
+export const useActionButtonLogic = ({
+	tokenId,
+	owned,
+	action,
+	onCannotPerformAction,
+}: UseActionButtonLogicProps) => {
+	const { address } = useAccount();
+	const actionsThatOwnersCannotPerform = [
+		CollectibleCardAction.BUY,
+		CollectibleCardAction.OFFER,
+	];
+	const pendingActionType = actionButtonStore.pendingAction.type.get();
+
+	// Handle owner restrictions
+	// biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
+	useEffect(() => {
+		if (
+			owned &&
+			actionButtonStore.pendingAction.get() &&
+			address &&
+			!actionsThatOwnersCannotPerform.includes(action) &&
+			actionButtonStore.pendingAction.get()?.collectibleId === tokenId
+		) {
+			onCannotPerformAction?.(
+				pendingActionType as
+					| CollectibleCardAction.BUY
+					| CollectibleCardAction.OFFER,
+			);
+		}
+	}, [
+		owned,
+		actionButtonStore.pendingAction.get(),
+		address,
+		action,
+		tokenId,
+		onCannotPerformAction,
+		pendingActionType,
+	]);
+
+	// Execute pending action when user becomes connected
+	useEffect(() => {
+		if (
+			address &&
+			!owned &&
+			actionButtonStore.pendingAction.get() &&
+			actionButtonStore.pendingAction.get()?.collectibleId === tokenId
+		) {
+			executePendingActionIfExists();
+		}
+	}, [address, owned, tokenId]);
+
+	const shouldShowAction = !address
+		? [CollectibleCardAction.BUY, CollectibleCardAction.OFFER].includes(action)
+		: true;
+
+	const isOwnerAction =
+		address &&
+		owned &&
+		[
+			CollectibleCardAction.LIST,
+			CollectibleCardAction.TRANSFER,
+			CollectibleCardAction.SELL,
+		].includes(action);
+
+	return {
+		address,
+		shouldShowAction,
+		isOwnerAction,
+	};
+};
